requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
asyncpg==0.30.0
beautifulsoup4==4.8.2
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cmudict==1.0.32
cobble==0.1.4
compressed-rtf==1.0.7
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
deprecation==2.1.0
docx2txt==0.9
ebcdic==1.1.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
extract-msg==0.28.7
eyeD3==0.9.8
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.1
fsspec==2025.5.1
greenlet==3.2.2
h11==0.16.0
hmmlearn==0.3.3
idna==3.10
IMAPClient==2.1.0
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
lxml==5.4.0
mammoth==1.9.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
murmurhash==1.0.13
narwhals==1.41.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
olefile==0.47
openai-whisper==20240930
packaging==25.0
pdfminer.six==20191110
pillow==11.2.1
plotly==6.1.2
preshed==3.0.10
psycopg2-binary==2.9.10
pyAudioAnalysis==0.3.14
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.26.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==0.6.23
regex==2024.11.6
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
shellingham==1.5.4
six==1.12.0
sklearn-compat==0.1.3
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
textract==1.6.5
textstat==0.7.7
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.0
torchaudio==2.7.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
xlrd==1.2.0
XlsxWriter==3.2.3
zipp==3.22.0
